src/VumbaSoft.AdventureWorks.Controllers/Demografic/Tenants/TenantsController.cs
using Microsoft.AspNetCore.Mvc;
using VumbaSoft.AdventureWorks.Objects;
using VumbaSoft.AdventureWorks.Services;
using VumbaSoft.AdventureWorks.Validators;
using System;

namespace VumbaSoft.AdventureWorks.Controllers.Demografic
{
    [Area("Demografic")]
    public class TenantsController : ValidatedController<ITenantValidator, ITenantService>
    {
        public TenantsController(ITenantValidator validator, ITenantService service)
            : base(validator, service)
        {
        }

        [HttpGet]
        public ViewResult Index()
        {
            return View(Service.GetViews());
        }

        [HttpGet]
        public ViewResult Create()
        {
            return View();
        }

        [HttpPost]
        public ActionResult Create(TenantView tenant)
        {
            if (!Validator.CanCreate(tenant))
                return View(tenant);

            Service.Create(tenant);

            return RedirectToAction("Index");
        }

        [HttpGet]
        public ActionResult Details(Int32 id)
        {
            return NotEmptyView(Service.Get<TenantView>(id));
        }

        [HttpGet]
        public ActionResult Edit(Int32 id)
        {
            return NotEmptyView(Service.Get<TenantView>(id));
        }

        [HttpPost]
        public ActionResult Edit(TenantView tenant)
        {
            if (!Validator.CanEdit(tenant))
                return View(tenant);

            Service.Edit(tenant);

            return RedirectToAction("Index");
        }

        [HttpGet]
        public ActionResult Delete(Int32 id)
        {
            return NotEmptyView(Service.Get<TenantView>(id));
        }

        [HttpPost]
        [ActionName("Delete")]
        public RedirectToActionResult DeleteConfirmed(Int32 id)
        {
            Service.Delete(id);

            return RedirectToAction("Index");
        }
    }
}
